Refactor text in hoerl.md for conciseness
Removed duplicate paragraph and adjusted wording for clarity. (#16)

* Add post on art reflection and societal themes

This post reflects on the contrast between two artworks by Seiwert and Hoerle, exploring themes of utopia and warning in the context of societal unrest during the Weimar Republic. It captures the protagonist Faulmann's thoughts and experiences in the museum and later at a Christmas market.

* Add image to 2025-12-01-hoerl post

Added an image to the post for visual enhancement.

* Refactor text in hoerl.md for conciseness

Removed duplicate paragraph and adjusted wording for clarity.