Cafeedt (#18)
* Create post about art and culture in café setting

Add a new post titled 'Zwischen Bühne und Tresen' discussing the intersection of art, music, and culture in a café setting.

* Revise content in 2025-12-12-cafee.md

Updated the text to enhance clarity and flow, including adjustments to the descriptions of the artwork and the relationship between the artists.

* add remark towards pinakothel version

* more on the cat scene

* no Schmidt

* improve location and srory

* viewer becmoses part of the picture

* cat mod

* more on punk

* Fixed editor versions

* Holzsäule extended

* Biermann

* Some footnotes

* Supporting hand

* Fix the hand part

* add images

* add another cover image

* add dachsbert style

* remove image

* reworking the whole text

* fixed markdown div

* Linked list

* add link to image

* small fix

* Epilog

* Delete artist recommendations and related links

Removed sections discussing various artists and their works, including Jörg Immendorff, A. R. Penck, and others, along with their related links and descriptions.